requirements.txt
Click==7.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
PyMySQL==1.0.2
PyYAML==5.3.1
SQLAlchemy==1.4.7
Werkzeug==1.0.1
zipp==1.0.0
